+++ src/Commands/Command_ChangeNodeDefinition.h
@@ -18,13 +18,13 @@ struct Command_ChangeNodeDefinition {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Node id", node_id),
-            cereal::make_nvp("New value", new_value)
+            ser20::make_nvp("Node id", node_id),
+            ser20::make_nvp("New value", new_value)
         );
     }
 };
@@ -40,13 +40,13 @@ struct ReversibleCommand_ChangeNodeDefinition {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Forward", fwd),
-            cereal::make_nvp("Old value", old_value)
+            ser20::make_nvp("Forward", fwd),
+            ser20::make_nvp("Old value", old_value)
         );
     }
 };