+++ src/Commands/Command_RemoveLink.h
@@ -18,13 +18,13 @@ struct Command_RemoveLink {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Link id", link_id),
-            cereal::make_nvp("Link", link)
+            ser20::make_nvp("Link id", link_id),
+            ser20::make_nvp("Link", link)
         );
     }
 };
@@ -39,12 +39,12 @@ struct ReversibleCommand_RemoveLink {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Forward", fwd)
+            ser20::make_nvp("Forward", fwd)
         );
     }
 };