+++ src/Commands/Command_RemoveNode.h
@@ -18,13 +18,13 @@ struct Command_RemoveNode {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Node id", node_id),
-            cereal::make_nvp("Node", node)
+            ser20::make_nvp("Node id", node_id),
+            ser20::make_nvp("Node", node)
         );
     }
 };
@@ -40,13 +40,13 @@ struct ReversibleCommand_RemoveNode {
 
 private:
     // Serialization
-    friend class cereal::access;
+    friend class ser20::access;
     template<class Archive>
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Forward", fwd),
-            cereal::make_nvp("Node position", node_pos)
+            ser20::make_nvp("Forward", fwd),
+            ser20::make_nvp("Node position", node_pos)
         );
     }
 };