+++ src/ProjectManager/Command_OpenProject.h
@@ -12,9 +12,9 @@ struct Command_OpenProject {
 
 } // namespace Lab
 
-namespace cereal {
+namespace ser20 {
 template<class Archive>
 void serialize(Archive&, Lab::Command_OpenProject&)
 {
 }
-} // namespace cereal
+} // namespace ser20